make infer work again
This required a change to the `successors` function, used by the topological sort. It must return only the successors defined in the local scope, otherwise
a unwrap in the graph code will trigger a panic. This is safe because symbols outside of the local scope (either further up or lower down) cannot form a cycle