prevent duplicate definitions
patterns that bind multiple symbols, like `{ x, y } = someDef` would insert the definition twice (once for x, once for y).
that can lead to problems in codegen, and means the work for this definition is done twice.

Now we use the region of the pattern as a key, to ensure the pattern and its associated expression is only processed once.