don't union vars_by_symbol
The symbols will be re-introduced as needed from the def_types. Besides, the vars_by_symbol share variables (so, multiple symbols from different modules can now point to the same variable in the current module's Subs). So this is both correct, and limits the number of items in the vars_by_symbol map to just the ones needed by the current module