Drop Num.endi
It's [not portable](https://doc.rust-lang.org/std/primitive.u32.html#method.to_ne_bytes),
and apparently (in a Rust discussion somewhere) it's been said that the
host endianness is very rarely useful in user code. Given all that, it seems
reasonable to leave it up to platforms to expose it (or not).