Support polymorphism immediates capturing other polymorphism immediates
This patch introduces the notion of a `ZeroArgumentClosure`, which has
all the semantics and behavior of a `ZeroArgumentThunk`, except that it
also closes over some number of variables.

`ZeroArgumentClosure`s can only ever be created during the demotion of a
symbol to a thunk, so they aren't present in the surface syntax.

This enables the specialization of `b` in expressions like

```
a = 1
b = A a
```

to all sorts of types including `[A U8]`, `[A U32]`, `[A U8 | B]`, etc.
This was previously not possible because specialization of `b` demands a
specialization of `a`, and moreover, may require the capture of other
variables that are already specialized.