Always deep-copy compound types that reference deep-copied types
This closes #2379. Here is where we usually give a small snippet of what
caused that issue and why this patch fixes it. For posterity and hints
on debugging issues like this, I'll now give a longer form of my process
here.

Okay, so our first step is to try to reduce the reproducer given in
https://github.com/rtfeldman/roc/issues/2379#issuecomment-1029084894
further. I was able to get to this:

```roc
app "hello-rust"
    packages { pf: "platform" }
    imports [  ]
    provides [ main ] to pf

Job : [ Job (List Job) [ A ] ]

job : Job

bundle = job

main = ""
```

```bash
$ cargo run check examples/hello-rust/Hello.roc
...
thread '<unnamed>' panicked at 'rank was 1 but recursion var 133 has higher rank 2', compiler/solve/src/solve.rs:1568:25
```

Okay. Now that it's in a single module we can but it in a "solve" test
easily.

```diff
--- a/compiler/solve/tests/solve_expr.rs
+++ b/compiler/solve/tests/solve_expr.rs
+
+    #[test]
+    fn issue_2379() {
+        infer_eq_without_problem(
+            indoc!(
+                r#"
+                Job : [ Job [ Command ] (List Job) ]
+
+                job : Job
+
+                bundle = job
+
+                bundle
+                "#
+            ),
+            "",
+        )
+    }
 }
```

```
$ cargo test -p roc_solve issue_2379 -- --nocapture

...

running 1 test
thread '<unnamed>' panicked at 'rank was 1 but recursion var 134 has higher rank 3', compiler/solve/src/solve.rs:1568:25
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

Huh.. the rank is higher (3) than what we saw before (2)! Looks like
this is might be a (de)generalization bug. Let's get rid of one
let-generalization.

```diff
--- a/compiler/solve/tests/solve_expr.rs
+++ b/compiler/solve/tests/solve_expr.rs
+
+    #[test]
+    fn issue_2379() {
+        infer_eq_without_problem(
+            indoc!(
+                r#"
+                Job : [ Job [ Command ] (List Job) ]
+
+                job : Job
+
+                job
+                "#
+            ),
+            "",
+        )
+    }
 }
```

```
$ cargo test -p roc_solve issue_2379 -- --nocapture

...

running 1 test
thread '<unnamed>' panicked at 'rank was 1 but recursion var 132 has higher rank 2', compiler/solve/src/solve.rs:1568:25
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

Alright, there we go. Now, let's observe definitions of `Job` that don't
induce this panic:

1. `Job : [ Job { a : Job } (List Job) ]`
2. `Job : [ Job [] (List Job) ]`
3. `Job : [ Job (List Job) ]`
4. `Job : [ Job [ Command ] { j: Job } ]`

I can't gleam a lot of hints out of the first three (they're actually
just a symptom, we'll get back to them later). But the fourth is
interesting. If we don't wrap `Job` in an alias, everything is fine.
Looks like we're probably just forgetting to update some variables when
copying aliases.

```rust
// fn deep_copy_var_help
        Alias(symbol, arguments, real_type_var) => {
            let new_variables =
                SubsSlice::reserve_into_subs(subs, arguments.all_variables_len as _);
            for (target_index, var_index) in (new_variables.indices()).zip(arguments.variables()) {
                let var = subs[var_index];
                let copy_var = deep_copy_var_help(subs, max_rank, pools, visited, var);
                subs.variables[target_index] = copy_var;
            }

            let new_arguments = AliasVariables {
                variables_start: new_variables.start,
                ..arguments
            };

            let new_real_type_var =
                deep_copy_var_help(subs, max_rank, pools, visited, real_type_var);
            let new_content = Alias(symbol, new_arguments, new_real_type_var);

            subs.set(copy, make_descriptor(new_content));

            copy
        }
```

Huh.. no, that seems fine. Okay. I still think it's a generalization
bug. So let's look at what's getting re-generalized and what's not.

Let's instrument some `debug`s in `deep_copy_var_help` (the function
responsible for, well, copying types to generalize them). After adding
them to the appropriate places we get a trace like the following:

```
...
[compiler/solve/src/solve.rs:1348] var = 137
[compiler/solve/src/solve.rs:1348] c = Alias(`#List.List`, [117, 134], 136)
[compiler/solve/src/solve.rs:1416] s = "137|2|Mark(3)|Mark(3)|Mark(4)"
...
[compiler/solve/src/solve.rs:1348] var = 139
[compiler/solve/src/solve.rs:1348] c = [ Global('Job') [137, 138], ]<2> as <117>
[compiler/solve/src/solve.rs:1416] s = "139|2|Mark(3)|Mark(3)|Mark(4)"
...
[compiler/solve/src/solve.rs:1818] var = 117
[compiler/solve/src/solve.rs:1818] copy = 144
...
[compiler/solve/src/solve.rs:291] rank = 2
[compiler/solve/src/solve.rs:291] var = 126
[compiler/solve/src/solve.rs:291] actual = 141
...
[compiler/solve/src/solve.rs:1431] var = 142
[compiler/solve/src/solve.rs:1431] roc_types::subs::SubsFmtContent(content, subs) = [ Global('Job') [123, 147], ]<143> as <144>
[compiler/solve/src/solve.rs:1416] s = "143|2|Mark(5)|Mark(5)|Mark(6)"
...
[compiler/solve/src/solve.rs:1570] var = 123
[compiler/solve/src/solve.rs:1570] roc_types::subs::SubsFmtContent(c, subs) = Alias(`#List.List`, [117, 120], 122)
[compiler/solve/src/solve.rs:1416] s = "123|1|Mark(4)|Mark(5)|Mark(6)"
...
[compiler/solve/src/solve.rs:1585] rank = 1
[compiler/solve/src/solve.rs:1585] rec_var_rank = 2
thread '<unnamed>' panicked at 'rank was 1 but recursion var 144 has higher rank 2', compiler/solve/src/solve.rs:1587:25
```

Aside: I want to create some better debugging tools for the typechecker!
A graphviz visualizer would be a good first step.

Now we can parse out what's happening. The generalized type of `job` is
`[ Job [ Command ] (List <117>) ] as 117`. But when we then use `job` as
the last expression, we must create a copy of this type that is
degeneralized. The trace shows us that the copy of the type is
`[ Job [ Command ] (List <117>) ] as 144` - the recursion variable `144`
is specialized, but the variable in the `List` alias is still
referencing generalized recursion variable! And that's the source of our
problems.

The reason the variable is not copied in the alias is because of these
branches in `deep_copy_var_help`:

```rust
    if let Some(copy) = desc.copy.into_variable() {
        return copy;
    } else if desc.rank != Rank::NONE {
        return var;
    }
```

`Rank::NONE` is reserved for variables that always need generalization.
Now `List <117>` doesn't have rank 0, even though `117` does, because
`List`s are aliases on the toplevel and hence have a rank of 1. So, if
we allow the copying of variables that reference variables that are
copied, that should fix the present issue:

```diff
--- a/compiler/solve/src/solve.rs
+++ b/compiler/solve/src/solve.rs
@@ -1795,7 +1795,7 @@ fn deep_copy_var_help(

     if let Some(copy) = desc.copy.into_variable() {
         return copy;
-    } else if desc.rank != Rank::NONE {
+    } else if desc.rank != Rank::NONE && !subs.var_contains(var, |d| d.copy.is_some()) {
         return var;
     }
```

Indeed it does.

And, now we have a better test title:

```diff
-   fn issue_2379() {
+   fn copy_vars_referencing_copied_vars() {
```