Advertise larger Roc project ideas in CONTRIBUTING
This document summarizes larger implementation- or research-heavy
projects that may be suitable for someone who wants to contribute to Roc
via a self-contained project. Most of these projects are expected to
require sustained effort, which also makes them possibly suitable for
academic theses.

Feel free to message me on Zulip to get edit rights to the document!