More efficient representation of args in the arg parser
Rather than taking slices of a list when parsing subcommands and the
like, we can mark certain positions in the arguments list as having
already been parsed. This allows us to keep the argument list constant,
at just the precise of an extra set (which is likely to be small in
practice).

This also makes it easy to show what arguments are leftover, if the time
comes for that. It also makes parsing positional arguments easy; you
just grab the next unmarked arguments.