Some attempts at detecting when type annotations are required
This is going to be much trickier than I anticipated. At the least, we
will need syntax-directed instructions on when to actually check for
flex able vars that are not resolvable, which should happen after
function body inference. It's not enough to do such checks every type we
check a def, since those types maybe escape into the nested scopes they
are defined in, and be found to be bound after checking.