Fix roc_error_macros::assert_sizeof_non_wasm! calls
It complained about the transmute failing due to different type sizes.
Tweaking the size calculations in this way fixed the build on my machine.

This failed on my M1 Macbook Pro running macOS 12.6.2

I am not sure if this is the right fix, but since it fixed the build
on my machine I decided to create this PR.