Make web REPL warning more accurate
The existing warning says there are "some features" missing,
and that "Missing Final Expression" is only an _example_.
But that makes it sound like it's generally unreliable and
could be missing any features, and nobody really knows.
The truth is that "Missing Final Expression" is the one and only
feature that's missing. There is no uncertainty about this.