Use reg from StorageManager instead of XR while loading args
When a function returns via arg pointer, the caller sets XR (x8) to
the address where the return value goes. We were sometimes using XR
as a temporary register while loading args, which lead to a segfault
when returning.

It will now request a general available register from StorageManager
instead.