Update the alignment of 128 bit numeric types
With current llvm and languages we work with (rust and zig), i128/u128/dec are not aligned to 16 bytes.

This commit sets their alignment to match what is found in zig and rust.
Once that is update, we can change back to 16 byte alignment for these.
That may take a while, it will at least be the next llvm release.
It is unclear when the changes will then be pulled into rust and zig.