Report import header as errors and format to new syntax
Whenever we encounter a header with an imports keyword, we will now
report it as an error and suggest the user to run `roc format`
which will automatically upgrade it to the new syntax.

This updates a lot of .roc files in the repo to use the new syntax
so that most test pass. Some are still failing (mostly in `roc_cli`),
but that seems to be a bug in canonicalization and not due to the
changes introduced in this commit.