Introduce new `module` header
Implements the new `module` header syntax as described in "module and package changes" [1]:

```
module [Request, Response, req]
```

Running `roc format` will automatically update the old `interface` headers
to the new syntax.

This commit updates a lot of .roc files in tests and examples to the new
syntax, but some are still missing. We'll do a final pass updating everything
when the other header types are updated to their new syntaxes too.

[1] https://docs.google.com/document/d/1E_77fO-44BtoBtXoVeWyGh1xN2KRTWTu8q6i25RNNx0/edit