Fix REPL header and execution mode
The REPL was specifying an ephemeral platform through the `app ... provides main to "./platform"`
syntax. This worked because in that case we wouldn't try to load any platforms, but now we do.

This commit introduces `ExecutionMode::ExecutableEval` which works like `Executable`, except that
it doesn't require a platform since it's only used to evaluate expressions. I used to to fix
the REPL, but I believe it could also be used for tests that were relying on the same behavior.