Add llvm generic_compare implementation for integers
This implements the following calculation posted by Richard on Zulip:
https://roc.zulipchat.com/#narrow/stream/304641-ideas/topic/ordering.2Fsorting.20ability/near/403858126

I'm using the compare.rs module that implements equality as a guide,
some the code has some similarities with that, but also I'm starting
with a blank page to not overwhelm myself, which might result in some
differences (at least at first).

The compare.rs module we might want to rename to eq.rs in a future
commit. That would free up the name compare.rs to implement generic_eq.