Fix up SmallStringInterner and use GPA
This PR ended up mixing two things kinda unintentionally.

Firstly, it improves the small string interner.
The interner uses an unmanaged hash map with context to store deduplicated strings.
Since the strings are all meant to be small, instead of storing lengths, it just null terminates the strings.
Also, removed some extra mappins that are not needed to be in the small string interner.
It only deals with storing unique small strings and giving them an index

As part of the above update, I changed the interner to use a general purpose allocator instead of an arena.

That change led to a much larger change of following the error trailing and updating many things from an arena to a general purpose allocator.
In general, roc will actually use very few arenas in the default path.
Arenas are only for allocations that never grow in size.
Since roc is storing everything very densely, almost everything is in some form of an arraylist.
These all regularly grow in size and are not good for arena use.