Remove `||` and `&&` from tokenizer
As discussed here: https://roc.zulipchat.com/#narrow/channel/395097-compiler-development/topic/zig.20compiler.20-.20spike/near/504579627

I assume in the long term, support will be added to the parser
to notice when double opBar or opAmpersand should be consider an `and` or `or`.
Then the formatter will fix things up.

For now, double bar is always a lambda and double ampersand is invalid.