Add experimental LSP written in Zig (#8423)
* Add transport layer for LSP

Takes care of handling the messages sent and received between the editor and the LSP. Also allows an optional log file to monitor the message transmitted while the stdin and stdout are taken control of.

* Add LSP cli command

Allows starting the LSP with `experimental-lsp` and allows debugging with the `--debug-transport` flag.

* Add the LSP module

* Add the server for the LSP

Tracks the state of the server and handle the request. At this point in time, only supports `initialized` and `exit` messages.

* Add LSP data structure definition

* Add a README.md

* Modernize ArrayList usage

Moved from array_list.Managed and ArrayListUnmanaged to ArrayList

* Fix risk of infinite loop

Protects against a case where writeSome returns 0 which would cause an infinite loop

* Fix typo and incorrect error handling

Unsaved duplicate line from previous commit was removed and the wrong error return was changed to a panic. This behavior is not supposed to happen unless the writeSome function was modified.

* Complete default InitializeResult

Returns an empty server capabilities struct to indicate where server capabilities will go in the future. Also provide a version in case null is mishandled by the editor. Furthermore, add logging on incorrect payload object.

* Handle default state more accurately

Instead of using `else` when handling shutdown, expect the state that should expect a shutdown.

* Abstract the LSP methods implementation

Allows the server to now wire and dispatch handlers that can be defined separately for every method thats implemented. Makes for cleaner code and prevent cluttering the `server.zig` file.

* Add test aggregator for all LSP related tests

* Change LSP log to single file

Now appends the log of the LSP session to a single file which makes it easier to watch it without changing file at every editor session restart.

* Add implementation instructions to README.md

* Fix typo

* Add doc comment to pub functions

* wire up tests

---------

Co-authored-by: Anton-4 <17049058+Anton-4@users.noreply.github.com>