Fix the imports not resolving sometimes
Because WaitingOnImports were only checked once, if they are not ready the build would continue as usual leaving them in limbo. They are now being checked later down the pipeline to be processed properly. This was essential for single threaded build which has no global queue so they were always forgotten